Use forward iterator for lookups; use end IP as key
The lookups now use a forward database iterator, which performs
significantly better than a reverse iterator. For this to work, the
LevelDB database now stores the end of the range as the key, and uses
the begin of the range for the "within range" test during lookup. (This
used to be the other way around). This also obsoletes the need to
call incr_ip() for each lookup, making it even more efficient.